+++ packages/vue/src/components/button/CmButton.vue
@@ -13,6 +13,7 @@ const props = withDefaults(defineProps<CmButtonProps>(), {
   size: 'md',
   type: 'button',
   disabled: false,
+  loading: false,
 });
 
 const attrs = useAttrs();
@@ -20,11 +21,20 @@ const attrs = useAttrs();
 const classes = computed(() =>
   mergeCmClasses('cm-button', `cm-button--${props.variant}`, `cm-button--${props.size}`, attrs.class as CmClassValue),
 );
-const rootAttrs = computed(() => omitCmOwnedAttrs(attrs, ['type', 'disabled']));
+const rootAttrs = computed(() => omitCmOwnedAttrs(attrs, ['type', 'disabled', 'aria-busy']));
 </script>
 
 <template>
-  <button v-bind="rootAttrs" :class="classes" :type="props.type" :disabled="props.disabled">
+  <button
+    v-bind="rootAttrs"
+    :class="classes"
+    :type="props.type"
+    :disabled="props.disabled || props.loading"
+    :aria-busy="props.loading ? 'true' : undefined"
+  >
+    <span v-if="props.loading" class="cm-button__spinner" aria-hidden="true" />
+    <span v-else-if="$slots.leading" class="cm-button__leading"><slot name="leading" /></span>
     <span class="cm-button__label"><slot /></span>
+    <span v-if="$slots.trailing" class="cm-button__trailing"><slot name="trailing" /></span>
   </button>
 </template>