+++ packages/vue/src/components/button/button.types.ts
@@ -9,4 +9,5 @@ export interface CmButtonProps {
   size?: CmButtonSize;
   type?: CmButtonType;
   disabled?: boolean;
+  loading?: boolean;
 }